Add Expense Button and Form Working and more.
1.) The add expense button now spawns a new window that allows the user
to enter the information needed to create a new expense. If the user
does not enter a category for the expense, it defaults to misc.

2.) The "Delete Expense" button is now working too. The user can select
an expense in the ListView and click the button to delete an expense.
If the user does not select an expense to delete but clicks the button,
the will be notified via an alert window that they need to select an
expense to delete.