Budget Stats Model Code and Binding Started
1.) Created a BudgetStats class that will be used to perform budget
calculations that are bindable. That way, when a property updates, it is
automatically recalculated.

2.) I have monthly income bound and the monthly expenses calculation
bound too. I found that the javafx.beans.binding package has a way to
create custom bindings so calculations can be performed inside bindings.
This occurs in BudgetStats.java.

TODO: Push this to remote repo and work on desktop.